requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
asyncpg==0.30.0
black==24.10.0
click==8.1.7
dill==0.3.9
fastapi==0.115.5
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==5.13.2
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.2.0
pylint==3.3.1
python-dotenv==1.0.1
ruff==0.8.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.32.1
